make async message visible, make deletion of async message customizable

+++ amqppeclconnector.php
@@ -58,10 +58,11 @@ class PECLAMQPConnector extends AbstractAMQPConnector
 	 * Return result of task execution for $task_id
 	 * @param AMQPConnection $connection Connection object
 	 * @param string $task_id Celery task identifier
+     * @param boolean $removeMessageFromQueue whether to remove message from queue
 	 * @return array array('body' => JSON-encoded message body, 'complete_result' => AMQPEnvelope object)
 	 * 			or false if result not ready yet
 	 */
-	function GetMessageBody($connection, $task_id)
+	function GetMessageBody($connection, $task_id, $removeMessageFromQueue = true)
 	{
 		$this->Connect($connection);
 		$ch = new AMQPChannel($connection);
@@ -75,7 +76,9 @@ class PECLAMQPConnector extends AbstractAMQPConnector
 		}
 		catch(AMQPQueueException $e)
 		{
-   			$q->delete();
+            if ($removeMessageFromQueue) {
+                $q->delete();
+            }
 			$connection->disconnect();
 			return false;
 		}
@@ -84,14 +87,18 @@ class PECLAMQPConnector extends AbstractAMQPConnector
 
 		if(!$message) 
 		{
-   			$q->delete();
+            if ($removeMessageFromQueue) {
+                $q->delete();
+            }
 			$connection->disconnect();
 			return false;
 		}
 
 		if($message->getContentType() != 'application/json')
 		{
-			$q->delete();
+            if ($removeMessageFromQueue) {
+                $q->delete();
+            }
 			$connection->disconnect();
 
 			throw new CeleryException('Response was not encoded using JSON - found ' . 
@@ -99,7 +106,9 @@ class PECLAMQPConnector extends AbstractAMQPConnector
 				' - check your CELERY_RESULT_SERIALIZER setting!');
 		}
 
-		$q->delete();
+        if ($removeMessageFromQueue) {
+            $q->delete();
+        }
 		$connection->disconnect();
 
 		return array(
